requirements-dev.txt
pyhamcrest
doublex
mamba
coverage==3.7
coveralls
